rename Helper.js in lowercase
