+++ src/components/TrendingFeedItem.js
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import { numberWithCommas } from '../utils/Helper';
+import { numberWithCommas } from '../utils/helper';
 import PropTypes from 'prop-types';
 
 const TrendingFeedItem = ({ trendingFeedServerData }) => {
